+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2020/05/22 Format Document Beta v0.3
+
+add switch case format
 
 ## 2020/04/24 add goto class Definition
 
@@ -7,4 +10,4 @@
 
 ## 2020/04/05 fork from
 
-- <https://github.com/cweijan/vscode-autohotkey>
+-   <https://github.com/cweijan/vscode-autohotkey>